update notification code :)

+++ LMS library/Controllers/QuestionController.cs
@@ -1,7 +1,10 @@
-﻿using LMS_library.Repositories;
+﻿using LMS_library.Data;
+using LMS_library.Migrations;
+using LMS_library.Repositories;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace LMS_library.Controllers
 {
@@ -12,11 +15,15 @@ namespace LMS_library.Controllers
     {
 
         private readonly IQuestionsRepository _repository;
+        private readonly INotificationRepository _notificationRepository;
+        private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly DataDBContex _contex;
-        public QuestionController(IQuestionsRepository repository, DataDBContex contex)
+        public QuestionController(INotificationRepository notificationRepository, IHttpContextAccessor httpContextAccessor, IQuestionsRepository repository, DataDBContex contex)
         {
             _repository = repository;
             _contex = contex;
+            _notificationRepository = notificationRepository;
+            _httpContextAccessor = httpContextAccessor;
         }
         [HttpGet("list")]
         public async Task<IActionResult> GetAll()
@@ -51,6 +58,7 @@ namespace LMS_library.Controllers
                 {
                     return BadRequest("Question already exists .");
                 }
+                await _notificationRepository.AddNotification($"Create question for course {model.courseName} successfully at {DateTime.Now.ToLocalTime()}", Int32.Parse(UserInfo()), false);
                 var newQuestion = await _repository.AddQuestionAsync(model);
                 return Ok(newQuestion);
             }
@@ -62,6 +70,9 @@ namespace LMS_library.Controllers
 
             try
             {
+                var question  = await _contex.Questions.FindAsync(id);
+                if (question == null) { return NotFound(); }    
+                await _notificationRepository.AddNotification($"Delete question of course {question.courseName} successfully at {DateTime.Now.ToLocalTime()}", Int32.Parse(UserInfo()), false);
                 await _repository.DeleteQuestionAsync(id);
                 return Ok("Delete Success !");
 
@@ -74,10 +85,12 @@ namespace LMS_library.Controllers
         {
             try
             {
-                if (model.id != id)
+                var question = await _contex.Questions.FindAsync(id);
+                if (model.id != id || question ==null)
                 {
                     return NotFound();
                 }
+                await _notificationRepository.AddNotification($"Edit question {question.questionName} successfully at {DateTime.Now.ToLocalTime()}", Int32.Parse(UserInfo()), false);
                 await _repository.UpdateQuestionAsync(id, model);
                 return Ok("Update Successfully");
             }
@@ -91,11 +104,16 @@ namespace LMS_library.Controllers
         {
             try
             {
-                if(questionQuantity !=easyQuestionQuantity+ nomalQuestionQuantity +hardQuestionQuantity )
+                var leader = await _contex.Users.Where(l => l.Role.name == "Leader").ToListAsync();
+                if (questionQuantity !=easyQuestionQuantity+ nomalQuestionQuantity +hardQuestionQuantity )
                 {
                     return BadRequest("Wrong question quantity!");
                 }
-
+                foreach (var l in leader)
+                {
+                    await _notificationRepository.AddNotification($"New exam file create for {courseName} at {DateTime.Now.ToLocalTime()},exam name is {examName} please approve/reject the file soon as you can !", l.id, false);
+                }
+                await _notificationRepository.AddNotification($"Create exam by existing questions on system successfully at {DateTime.Now.ToLocalTime()} .Exam name is {examName}", Int32.Parse(UserInfo()), false);
                 await _repository.AddExamByExistQuestion(courseName,examName,time,examNumber,pointRange,questionQuantity,easyQuestionQuantity,nomalQuestionQuantity,hardQuestionQuantity);
                 return Ok("Create Successfully");
             }
@@ -104,5 +122,12 @@ namespace LMS_library.Controllers
                 return BadRequest();
             }
         }
+
+        private string UserInfo()
+        {
+            var result = _httpContextAccessor!.HttpContext!.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return result;
+        }
+
     }
 }
